Réorganisation de l'architecture pour se conformer aux bonnes pratiques de next.JS
