+++ app/fonts.css
@@ -1,0 +1,5 @@
+@font-face {
+    font-family: 'Poppins';
+    src: url('../public/fonts/Poppins/Poppins-Regular.ttf') format('truetype');
+    /* Ajoutez ici d'autres variantes ou formats de la police si nécessaire */
+  }
