+++ app/layout.tsx
@@ -1,9 +1,9 @@
-import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
-import './globals.css'
-import baseStyle from './baseStyle.module.css';
+// layout.tsx
 
-const inter = Inter({ subsets: ['latin'] })
+import type { Metadata } from 'next';
+import './globals.css';
+import Header from '../components/header';
+import Footer from '../components/footer';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -13,12 +13,15 @@ export const metadata: Metadata = {
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
-      <body className={inter.className + ' .test'}>{children}</body>
-
-    </html>
-  )
+    <>
+      <Header />
+      <div>
+        {children}
+      </div>
+      <Footer />
+    </>
+  );
 }
