+++ components/coverSearchDetail.tsx
