+++ pages/account/page.tsx
@@ -1,17 +1,12 @@
 import React from "react";
-import Header from "../components/header";
-import Footer from "../components/footer";
 
 export default function accountPage() {
 return (
-
     <div>
-    <Header />
         <h1>Mon Compte</h1>
         <p>
             Cras ac fermentum nibh, quis feugiat tortor. Fusce rhoncus diam a molestie luctus. Fusce tempor ante quis urna elementum, eu tristique massa cursus. In et tortor odio. Ut nec diam et velit placerat feugiat vel vel purus.
         </p>
-    <Footer />
     </div>
 )
 }
