+++ pages/login/index.tsx
@@ -1,15 +1,12 @@
 import React from "react";
-import Header from "../components/header";
-import Footer from "../components/footer";
+
+import Card from '../../components/card';
 
 export default function loginPage() {
 return (
-
     <div>
-    <Header />
         <h1>LOGIN</h1>
         <p>Page de loggin.</p>
-    <Footer />
     </div>
 )
 }
