+++ pages/messenger/index.tsx
@@ -1,17 +1,13 @@
 import React from "react";
-import Header from "../components/header";
-import Footer from "../components/footer";
+import Card from '../../components/card';
 
 export default function messengerPage() {
 return (
-
     <div>
-    <Header />
         <h1>MESSAGERIE</h1>
         <p>
             In in placerat eros. Sed a sem rhoncus, gravida nisi in, ornare tellus. Sed urna mi, ullamcorper ut libero sed, ultricies dictum lacus. Cras ac fermentum nibh, quis feugiat tortor. Fusce rhoncus diam a molestie luctus. Fusce tempor ante quis urna elementum, eu tristique massa cursus. In et tortor odio. Ut nec diam et velit placerat feugiat vel vel purus.
         </p>
-    <Footer />
     </div>
 )
 }
