+++ pages/register/index.tsx
@@ -1,15 +1,11 @@
 import React from "react";
-import Header from "../components/header";
-import Footer from "../components/footer";
+import Card from '../../components/card';
 
 export default function registerPage() {
 return (
-
     <div>
-    <Header />
         <h1>REGISTER</h1>
         <p>Page d'inscription.</p>
-    <Footer />
     </div>
 )
 }
